move download_flan into mttl/dataloader/flan_utils.py
