+++ tests/test_misc.py
@@ -1,4 +1,4 @@
-from mttl.cli.dataset_create import download_flan
+from mttl.dataloader.flan_utils import download_flan
 
 
 def test_download_slice():
